+++ schedulerx3-20240624/src/Models/CreateJobRequest.php
@@ -96,6 +96,11 @@ class CreateJobRequest extends Model
     public $routeStrategy;
 
     /**
+     * @var string
+     */
+    public $script;
+
+    /**
      * @var int
      */
     public $startTime;
@@ -137,6 +142,7 @@ class CreateJobRequest extends Model
         'parameters' => 'Parameters',
         'priority' => 'Priority',
         'routeStrategy' => 'RouteStrategy',
+        'script' => 'Script',
         'startTime' => 'StartTime',
         'status' => 'Status',
         'timeExpression' => 'TimeExpression',
@@ -230,6 +236,10 @@ class CreateJobRequest extends Model
 
         if (null !== $this->routeStrategy) {
             $res['RouteStrategy'] = $this->routeStrategy;
+        }
+
+        if (null !== $this->script) {
+            $res['Script'] = $this->script;
         }
 
         if (null !== $this->startTime) {
@@ -335,6 +345,10 @@ class CreateJobRequest extends Model
 
         if (isset($map['RouteStrategy'])) {
             $model->routeStrategy = $map['RouteStrategy'];
+        }
+
+        if (isset($map['Script'])) {
+            $model->script = $map['Script'];
         }
 
         if (isset($map['StartTime'])) {