+++ schedulerx3-20240624/src/Models/GetJobExecutionResponseBody/data.php
@@ -1,0 +1,300 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\SchedulerX3\V20240624\Models\GetJobExecutionResponseBody;
+
+use AlibabaCloud\Dara\Model;
+
+class data extends Model
+{
+    /**
+     * @var string
+     */
+    public $appName;
+
+    /**
+     * @var int
+     */
+    public $attempt;
+
+    /**
+     * @var string
+     */
+    public $dataTime;
+
+    /**
+     * @var int
+     */
+    public $duration;
+
+    /**
+     * @var string
+     */
+    public $endTime;
+
+    /**
+     * @var string
+     */
+    public $executor;
+
+    /**
+     * @var string
+     */
+    public $jobExecutionId;
+
+    /**
+     * @var int
+     */
+    public $jobId;
+
+    /**
+     * @var string
+     */
+    public $jobName;
+
+    /**
+     * @var string
+     */
+    public $jobType;
+
+    /**
+     * @var string
+     */
+    public $parameters;
+
+    /**
+     * @var string
+     */
+    public $result;
+
+    /**
+     * @var int
+     */
+    public $routeStrategy;
+
+    /**
+     * @var string
+     */
+    public $scheduleTime;
+
+    /**
+     * @var string
+     */
+    public $serverIp;
+
+    /**
+     * @var string
+     */
+    public $startTime;
+
+    /**
+     * @var int
+     */
+    public $status;
+
+    /**
+     * @var int
+     */
+    public $timeType;
+
+    /**
+     * @var int
+     */
+    public $triggerType;
+    protected $_name = [
+        'appName' => 'AppName',
+        'attempt' => 'Attempt',
+        'dataTime' => 'DataTime',
+        'duration' => 'Duration',
+        'endTime' => 'EndTime',
+        'executor' => 'Executor',
+        'jobExecutionId' => 'JobExecutionId',
+        'jobId' => 'JobId',
+        'jobName' => 'JobName',
+        'jobType' => 'JobType',
+        'parameters' => 'Parameters',
+        'result' => 'Result',
+        'routeStrategy' => 'RouteStrategy',
+        'scheduleTime' => 'ScheduleTime',
+        'serverIp' => 'ServerIp',
+        'startTime' => 'StartTime',
+        'status' => 'Status',
+        'timeType' => 'TimeType',
+        'triggerType' => 'TriggerType',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->appName) {
+            $res['AppName'] = $this->appName;
+        }
+
+        if (null !== $this->attempt) {
+            $res['Attempt'] = $this->attempt;
+        }
+
+        if (null !== $this->dataTime) {
+            $res['DataTime'] = $this->dataTime;
+        }
+
+        if (null !== $this->duration) {
+            $res['Duration'] = $this->duration;
+        }
+
+        if (null !== $this->endTime) {
+            $res['EndTime'] = $this->endTime;
+        }
+
+        if (null !== $this->executor) {
+            $res['Executor'] = $this->executor;
+        }
+
+        if (null !== $this->jobExecutionId) {
+            $res['JobExecutionId'] = $this->jobExecutionId;
+        }
+
+        if (null !== $this->jobId) {
+            $res['JobId'] = $this->jobId;
+        }
+
+        if (null !== $this->jobName) {
+            $res['JobName'] = $this->jobName;
+        }
+
+        if (null !== $this->jobType) {
+            $res['JobType'] = $this->jobType;
+        }
+
+        if (null !== $this->parameters) {
+            $res['Parameters'] = $this->parameters;
+        }
+
+        if (null !== $this->result) {
+            $res['Result'] = $this->result;
+        }
+
+        if (null !== $this->routeStrategy) {
+            $res['RouteStrategy'] = $this->routeStrategy;
+        }
+
+        if (null !== $this->scheduleTime) {
+            $res['ScheduleTime'] = $this->scheduleTime;
+        }
+
+        if (null !== $this->serverIp) {
+            $res['ServerIp'] = $this->serverIp;
+        }
+
+        if (null !== $this->startTime) {
+            $res['StartTime'] = $this->startTime;
+        }
+
+        if (null !== $this->status) {
+            $res['Status'] = $this->status;
+        }
+
+        if (null !== $this->timeType) {
+            $res['TimeType'] = $this->timeType;
+        }
+
+        if (null !== $this->triggerType) {
+            $res['TriggerType'] = $this->triggerType;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['AppName'])) {
+            $model->appName = $map['AppName'];
+        }
+
+        if (isset($map['Attempt'])) {
+            $model->attempt = $map['Attempt'];
+        }
+
+        if (isset($map['DataTime'])) {
+            $model->dataTime = $map['DataTime'];
+        }
+
+        if (isset($map['Duration'])) {
+            $model->duration = $map['Duration'];
+        }
+
+        if (isset($map['EndTime'])) {
+            $model->endTime = $map['EndTime'];
+        }
+
+        if (isset($map['Executor'])) {
+            $model->executor = $map['Executor'];
+        }
+
+        if (isset($map['JobExecutionId'])) {
+            $model->jobExecutionId = $map['JobExecutionId'];
+        }
+
+        if (isset($map['JobId'])) {
+            $model->jobId = $map['JobId'];
+        }
+
+        if (isset($map['JobName'])) {
+            $model->jobName = $map['JobName'];
+        }
+
+        if (isset($map['JobType'])) {
+            $model->jobType = $map['JobType'];
+        }
+
+        if (isset($map['Parameters'])) {
+            $model->parameters = $map['Parameters'];
+        }
+
+        if (isset($map['Result'])) {
+            $model->result = $map['Result'];
+        }
+
+        if (isset($map['RouteStrategy'])) {
+            $model->routeStrategy = $map['RouteStrategy'];
+        }
+
+        if (isset($map['ScheduleTime'])) {
+            $model->scheduleTime = $map['ScheduleTime'];
+        }
+
+        if (isset($map['ServerIp'])) {
+            $model->serverIp = $map['ServerIp'];
+        }
+
+        if (isset($map['StartTime'])) {
+            $model->startTime = $map['StartTime'];
+        }
+
+        if (isset($map['Status'])) {
+            $model->status = $map['Status'];
+        }
+
+        if (isset($map['TimeType'])) {
+            $model->timeType = $map['TimeType'];
+        }
+
+        if (isset($map['TriggerType'])) {
+            $model->triggerType = $map['TriggerType'];
+        }
+
+        return $model;
+    }
+}