+++ schedulerx3-20240624/src/Models/ListJobsResponseBody/data/records.php
@@ -124,6 +124,11 @@ class records extends Model
     public $routeStrategy;
 
     /**
+     * @var string
+     */
+    public $script;
+
+    /**
      * @var int
      */
     public $status;
@@ -181,6 +186,7 @@ class records extends Model
         'parameters' => 'Parameters',
         'priority' => 'Priority',
         'routeStrategy' => 'RouteStrategy',
+        'script' => 'Script',
         'status' => 'Status',
         'timeExpression' => 'TimeExpression',
         'timeType' => 'TimeType',
@@ -288,6 +294,10 @@ class records extends Model
 
         if (null !== $this->routeStrategy) {
             $res['RouteStrategy'] = $this->routeStrategy;
+        }
+
+        if (null !== $this->script) {
+            $res['Script'] = $this->script;
         }
 
         if (null !== $this->status) {
@@ -419,6 +429,10 @@ class records extends Model
 
         if (isset($map['RouteStrategy'])) {
             $model->routeStrategy = $map['RouteStrategy'];
+        }
+
+        if (isset($map['Script'])) {
+            $model->script = $map['Script'];
         }
 
         if (isset($map['Status'])) {