+++ schedulerx3-20240624/src/Models/UpdateJobShrinkRequest.php
@@ -94,6 +94,11 @@ class UpdateJobShrinkRequest extends Model
     public $routeStrategy;
 
     /**
+     * @var string
+     */
+    public $script;
+
+    /**
      * @var int
      */
     public $startTime;
@@ -130,6 +135,7 @@ class UpdateJobShrinkRequest extends Model
         'parameters' => 'Parameters',
         'priority' => 'Priority',
         'routeStrategy' => 'RouteStrategy',
+        'script' => 'Script',
         'startTime' => 'StartTime',
         'timeExpression' => 'TimeExpression',
         'timeType' => 'TimeType',
@@ -210,6 +216,10 @@ class UpdateJobShrinkRequest extends Model
 
         if (null !== $this->routeStrategy) {
             $res['RouteStrategy'] = $this->routeStrategy;
+        }
+
+        if (null !== $this->script) {
+            $res['Script'] = $this->script;
         }
 
         if (null !== $this->startTime) {
@@ -305,6 +315,10 @@ class UpdateJobShrinkRequest extends Model
 
         if (isset($map['RouteStrategy'])) {
             $model->routeStrategy = $map['RouteStrategy'];
+        }
+
+        if (isset($map['Script'])) {
+            $model->script = $map['Script'];
         }
 
         if (isset($map['StartTime'])) {